feat(TypeComposer): add new method `addNestedFields()`
This is the beginning of new graphql-compose' killer feature. For now it
just allows to use dots in fields names. But in the future it will be
used for simple but very convenient Authorization. Stay tuned!

Method `addNestedFields()` designed for accepting fieldnames with dots.
Under the hood it will create or reuse existed GraphQLObjectTypes. As
the result you will obtain well nested mutations.
```js
MutationTC.addNestedFields({
  'crm.page.create': PageTC.getResolver('createOne'),
  'crm.page.update': PageTC.getResolver('updateById'),
  'crm.page.remove': PageTC.getResolver('removeById'),

  'crm.admin.create': AdminTC.getResolver('createOne'),
  'crm.admin.update': AdminTC.getResolver('updateById'),
  'crm.admin.remove': AdminTC.getResolver('removeById'),
});
```